fix: keep cross-platform installer tests runnable on Windows
Only TestExtract depends on syscall.Umask, so move it into a new !windows-tagged http_installer_unix_test.go instead of tagging the whole file. This restores Windows coverage for NewForSource, isRemoteHTTPArchive, cleanJoin, mediaTypeToExtension, and the nested-extraction test, which the file-level build tag had excluded.
